revised formats

+++ main/data_preprocessing.py
@@ -5,22 +5,26 @@ import warnings
 import json
 import logging
 
-warnings.filterwarnings('ignore')
+warnings.filterwarnings("ignore")
 
 # Configure logging
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
 )
 
+
 # Function to load configuration settings from a JSON file
 def load_config(config_file_path):
-    with open(config_file_path, 'r') as file:
+    with open(config_file_path, "r") as file:
         config = json.load(file)
     return config
 
 
 # Configure logging
-logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
+logging.basicConfig(
+    level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
+)
+
 
 def preprocess_employee_data(df):
     """
@@ -41,36 +45,36 @@ def preprocess_employee_data(df):
         # Reset index to use it as Employee_ID and adjust it to start from 1
         logging.info("Resetting index and setting Employee_ID.")
         df.reset_index(inplace=True)
-        df.rename(columns={'index': 'Employee_ID'}, inplace=True)
-        df['Employee_ID'] = df['Employee_ID'] + 1
-        df['Employee_ID'] = df['Employee_ID'].astype(int)
+        df.rename(columns={"index": "Employee_ID"}, inplace=True)
+        df["Employee_ID"] = df["Employee_ID"] + 1
+        df["Employee_ID"] = df["Employee_ID"].astype(int)
 
         # Convert satisfaction levels and evaluations to floats
         logging.info("Converting satisfaction_level and last_evaluation to float.")
-        df['satisfaction_level'] = df['satisfaction_level'].astype(float)
-        df['last_evaluation'] = df['last_evaluation'].astype(float)
+        df["satisfaction_level"] = df["satisfaction_level"].astype(float)
+        df["last_evaluation"] = df["last_evaluation"].astype(float)
 
         # Convert counts and hours to integers
         logging.info("Converting project counts, hours, and time spent to integers.")
-        df['number_project'] = df['number_project'].astype(int)
-        df['average_monthly_hours'] = df['average_monthly_hours'].astype(int)
-        df['time_spend_company'] = df['time_spend_company'].astype(int)
+        df["number_project"] = df["number_project"].astype(int)
+        df["average_monthly_hours"] = df["average_monthly_hours"].astype(int)
+        df["time_spend_company"] = df["time_spend_company"].astype(int)
 
         # Convert binary columns to integers
         logging.info("Converting binary columns to integers.")
-        df['work_accident'] = df['work_accident'].astype(int)
-        df['left'] = df['left'].astype(int)
-        df['promotion_last_5years'] = df['promotion_last_5years'].astype(int)
+        df["work_accident"] = df["work_accident"].astype(int)
+        df["left"] = df["left"].astype(int)
+        df["promotion_last_5years"] = df["promotion_last_5years"].astype(int)
 
         # Convert categorical columns to 'category' data type
         logging.info("Converting department and salary columns to 'category' type.")
-        df['department'] = df['department'].astype('category')
-        df['salary'] = df['salary'].astype('category')
+        df["department"] = df["department"].astype("category")
+        df["salary"] = df["salary"].astype("category")
 
         # Convert encoded columns to integers
-        #logging.info("Converting salary_encoded and department_encoded to integers.")
-        #df['salary_encoded'] = df['salary_encoded'].astype(int)
-        #df['department_encoded'] = df['department_encoded'].astype(int)
+        # logging.info("Converting salary_encoded and department_encoded to integers.")
+        # df['salary_encoded'] = df['salary_encoded'].astype(int)
+        # df['department_encoded'] = df['department_encoded'].astype(int)
 
         logging.info("Data preprocessing completed successfully.")
         return df
@@ -86,7 +90,6 @@ def preprocess_employee_data(df):
     except Exception as e:
         logging.error(f"Unexpected error during preprocessing: {e}")
         raise RuntimeError(f"Unexpected error during preprocessing: {e}")
-
 
 
 def remove_outliers(data, z_score_threshold=3, numerical_columns=None):
@@ -107,17 +110,29 @@ def remove_outliers(data, z_score_threshold=3, numerical_columns=None):
     try:
         if isinstance(data, pd.DataFrame):
             if numerical_columns is None:
-                numerical_columns = ['satisfaction_level', 'last_evaluation', 'number_project',
-                                     'average_monthly_hours', 'time_spend_company', 'work_accident',
-                                     'promotion_last_5years']
+                numerical_columns = [
+                    "satisfaction_level",
+                    "last_evaluation",
+                    "number_project",
+                    "average_monthly_hours",
+                    "time_spend_company",
+                    "work_accident",
+                    "promotion_last_5years",
+                ]
 
             # Check which columns are present in the dataframe
-            available_columns = [col for col in numerical_columns if col in data.columns]
+            available_columns = [
+                col for col in numerical_columns if col in data.columns
+            ]
 
             if not available_columns:
-                raise ValueError(f"None of the specified numerical columns are available in the dataset: {numerical_columns}")
+                raise ValueError(
+                    f"None of the specified numerical columns are available in the dataset: {numerical_columns}"
+                )
 
-            logging.info(f"Processing available columns for outlier removal: {available_columns}")
+            logging.info(
+                f"Processing available columns for outlier removal: {available_columns}"
+            )
 
             # Calculating Z-scores for the available numerical columns
             z_scores = data[available_columns].apply(zscore)
@@ -127,8 +142,12 @@ def remove_outliers(data, z_score_threshold=3, numerical_columns=None):
 
             # Creating a new dataframe without outliers
             data_cleaned = data[~outliers]
-            logging.info(f"Outliers removed using Z-score threshold: {z_score_threshold}")
-            logging.info(f"Original data size: {data.shape[0]}, Cleaned data size: {data_cleaned.shape[0]}")
+            logging.info(
+                f"Outliers removed using Z-score threshold: {z_score_threshold}"
+            )
+            logging.info(
+                f"Original data size: {data.shape[0]}, Cleaned data size: {data_cleaned.shape[0]}"
+            )
 
             return data_cleaned
 
@@ -141,8 +160,12 @@ def remove_outliers(data, z_score_threshold=3, numerical_columns=None):
 
             # Returning a NumPy array without outliers
             data_cleaned = data[~outliers]
-            logging.info(f"Outliers removed using Z-score threshold: {z_score_threshold}")
-            logging.info(f"Original data size: {data.shape[0]}, Cleaned data size: {data_cleaned.shape[0]}")
+            logging.info(
+                f"Outliers removed using Z-score threshold: {z_score_threshold}"
+            )
+            logging.info(
+                f"Original data size: {data.shape[0]}, Cleaned data size: {data_cleaned.shape[0]}"
+            )
 
             return data_cleaned
